+++ duckscript_sdk/src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod annotation;
 pub(crate) mod condition;
 pub(crate) mod eval;
 pub(crate) mod exec;
@@ -5,4 +6,5 @@ pub(crate) mod flags;
 pub(crate) mod instruction_query;
 pub(crate) mod io;
 pub(crate) mod pckg;
+pub(crate) mod scope;
 pub(crate) mod state;